+++ domain/src/main/kotlin/com/ikkun2501/bookmanagement/domain/BookRepository.kt
@@ -5,8 +5,8 @@ package com.ikkun2501.bookmanagement.domain
  *
  */
 interface BookRepository {
-    fun findById(bookId: Long): Book?
+    fun findById(bookId: Int): Book?
     fun create(book: Book): Book
     fun update(book: Book): Book
-    fun delete(bookId: Long)
+    fun delete(bookId: Int)
 }
