+++ domain/src/main/kotlin/com/ikkun2501/bookmanagement/usecase/query/book/BookDetail.kt
@@ -11,10 +11,10 @@ package com.ikkun2501.bookmanagement.usecase.query.book
  * @property authorDescription
  */
 data class BookDetail(
-    val bookId: Long,
+    val bookId: Int,
     val title: String,
     val bookDescription: String,
-    val authorId: Long,
+    val authorId: Int,
     val authorName: String,
     val authorDescription: String
 )
