+++ domain/src/main/kotlin/com/ikkun2501/bookmanagement/usecase/query/book/BookQuery.kt
@@ -4,6 +4,6 @@ package com.ikkun2501.bookmanagement.usecase.query.book
  * 書籍参照系ユースケースのインターフェイス
  */
 interface BookQuery {
-    fun detail(bookId: Long): BookDetail
+    fun detail(bookId: Int): BookDetail
     fun search(bookSearchParams: BookSearchParams): List<BookSearchResultRow>
 }
